+++ src/BlurView.dom.js
@@ -1,3 +1,0 @@
-import BlurView from '../../expo/build/effects/BlurView.web';
-
-export default BlurView;