+++ src/NativeLinearGradient.dom.js
@@ -1,3 +1,0 @@
-import NativeLinearGradient from '../../expo/build/effects/NativeLinearGradient.web';
-
-export default NativeLinearGradient;