+++ src/facebook-ads.dom.js
@@ -1,3 +1,0 @@
-import * as facebookAds from '../../expo/build/facebook-ads/index.web';
-
-export default facebookAds;